Update List Challenge.py
Added comment to 9. Append Size
Hint by CodeAcademy to do it a different way
Use the range function to create a list of the numbers that you want to append to lst. Remember the range function can take two parameters. The first is the number you want the range to start at (we want 1). The second is the number you want the range to end at (exclusive). That number should relate to the length of lst.
Append every number in that range to lst by using append.